fix(jupiter): adjust breath placements and add star markers
- Bar 2: change breath from long to short
- Add star markers (🌠) after long breaths at Bar 5, 9, 21

+++ TsukiSound/Core/Audio/Synthesis/PureTone/Jupiter/JupiterMelodyData.swift
@@ -114,7 +114,7 @@ enum JupiterMelodyData {
 
         // === Bar 2 === ラ(4分) ラドシ.ソ(8+8+付点8+16)
         // ラ=1拍, ラ=0.5, ド=0.5, シ.=0.75, ソ=0.25 → 合計3拍 ✓
-        JupiterMelodyNote(.A4, bar: 2, beat: 0.0, dur: .quarter, breath: .long),     // ラ 🫧
+        JupiterMelodyNote(.A4, bar: 2, beat: 0.0, dur: .quarter, breath: .short),    // ラ 🫧
         JupiterMelodyNote(.A4, bar: 2, beat: 1.0, dur: .eighth),                     // ラ
         JupiterMelodyNote(.C5, bar: 2, beat: 1.5, dur: .eighth),                     // ド
         JupiterMelodyNote(.B4, bar: 2, beat: 2.0, dur: .dottedEighth),               // シ
@@ -137,6 +137,7 @@ enum JupiterMelodyData {
         // === Bar 5 === ミ(2分) ミソ
         // ミ(2分) ミ(8分)ソ(8分) → 2+0.5+0.5 = 3拍 ✓
         JupiterMelodyNote(.E4, bar: 5, beat: 0.0, dur: .half, breath: .long),     // ミ 🫧🫧
+        // 🌠
         JupiterMelodyNote(.E4, bar: 5, beat: 2.0, dur: .eighth),                  // ミ
         JupiterMelodyNote(.G4, bar: 5, beat: 2.5, dur: .eighth, breath: .short),  // ソ 🫧
 
@@ -164,6 +165,7 @@ enum JupiterMelodyData {
         // === Bar 9 === ド(2分) ソミ
         // ド(2分) ソ(8分)ミ(8分) → 2+0.5+0.5 = 3拍 ✓
         JupiterMelodyNote(.C5, bar: 9, beat: 0.0, dur: .half, breath: .long), // ド 🫧🫧
+        // 🌠
         JupiterMelodyNote(.G5, bar: 9, beat: 2.0, dur: .eighth),              // ソ
         JupiterMelodyNote(.E5, bar: 9, beat: 2.5, dur: .eighth),              // ミ
 
@@ -237,6 +239,7 @@ enum JupiterMelodyData {
 
         // === Bar 21 === ミ(2分) ミソ (Bar 5と同じパターン)
         JupiterMelodyNote(.E5, bar: 21, beat: 0.0, dur: .half, breath: .long),     // ミ 🫧🫧
+        // 🌠
         JupiterMelodyNote(.E5, bar: 21, beat: 2.0, dur: .eighth),                  // ミ
         JupiterMelodyNote(.G5, bar: 21, beat: 2.5, dur: .eighth, breath: .short),  // ソ 🫧
